# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для управления байками",
        "title": "Bike Microservice API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
        "/bikes": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создание нового байка",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bikes"
                ],
                "summary": "Создать байк",
                "parameters": [
                    {
                        "description": "Данные байка",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_handler_http.BikeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Байк создан",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_adapter_handler_http.successResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/webike_bike_microservice_nikita_internal_core_domain.Bike"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_handler_http.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_handler_http.errorResponse"
                        }
                    }
                }
            }
        },
        "/bikes/my": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получение всех байков авторизованного пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bikes"
                ],
                "summary": "Получить байки пользователя по айди пользователя",
                "responses": {
                    "200": {
                        "description": "Список байков пользователя",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_adapter_handler_http.successResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/webike_bike_microservice_nikita_internal_core_domain.Bike"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_handler_http.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_handler_http.errorResponse"
                        }
                    }
                }
            }
        },
        "/bikes/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получение информации о байке по ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bikes"
                ],
                "summary": "Получить байк",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID байка",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Байк найден",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_adapter_handler_http.successResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/webike_bike_microservice_nikita_internal_core_domain.Bike"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_handler_http.errorResponse"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_handler_http.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Байк не найден",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_handler_http.errorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновление данных байка",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bikes"
                ],
                "summary": "Обновить байк",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID байка",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_handler_http.UpdateBike"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Байк обновлен",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_adapter_handler_http.successResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/webike_bike_microservice_nikita_internal_core_domain.Bike"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_handler_http.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_handler_http.errorResponse"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_handler_http.errorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удаление байка",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bikes"
                ],
                "summary": "Удалить байк",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID байка",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Байк удален",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_handler_http.successResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_handler_http.errorResponse"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_handler_http.errorResponse"
                        }
                    }
                }
            }
        },
        "/bikes/{id}/with-components": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получение байка со всеми компонентами",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bikes"
                ],
                "summary": "Получить байк с компонентами",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID байка",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Байк с компонентами",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_adapter_handler_http.successResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/webike_bike_microservice_nikita_internal_core_domain.Bike"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_handler_http.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Байк не найден",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_handler_http.errorResponse"
                        }
                    }
                }
            }
        },
        "/bikes/{id}/with-user": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получение информации о байке и его владельце",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bikes"
                ],
                "summary": "Получить байк с пользователем",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID байка",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Байк с пользователем",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_adapter_handler_http.successResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_adapter_handler_http.BikeWithUserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_handler_http.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Байк не найден",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_handler_http.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "internal_adapter_handler_http.BikeRequest": {
            "type": "object",
            "required": [
                "mileage",
                "model",
                "type"
            ],
            "properties": {
                "mileage": {
                    "type": "integer",
                    "example": 1500
                },
                "model": {
                    "type": "string",
                    "example": "Mountain Bike Pro"
                },
                "type": {
                    "type": "string",
                    "example": "mountain"
                }
            }
        },
        "internal_adapter_handler_http.BikeWithUserResponse": {
            "type": "object",
            "properties": {
                "bike_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "mileage": {
                    "type": "integer",
                    "example": 1500
                },
                "model": {
                    "type": "string",
                    "example": "Mountain Bike Pro"
                },
                "user": {
                    "$ref": "#/definitions/internal_adapter_handler_http.User"
                }
            }
        },
...
        "internal_adapter_handler_http.UpdateBike": {
            "type": "object",
            "properties": {
                "mileage": {
                    "type": "integer",
                    "example": 2000
                },
                "model": {
                    "type": "string",
                    "example": "New Model"
                },
                "type": {
                    "type": "string",
                    "example": "mountain"
                }
            }
        },
...
        "internal_adapter_handler_http.User": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Иван Иванов"
                },
                "user_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                }
            }
        },
        "internal_adapter_handler_http.errorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Error"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "internal_adapter_handler_http.successResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string",
                    "example": "Success message"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "webike_bike_microservice_nikita_internal_core_domain.Bike": {
            "type": "object",
            "properties": {
                "bike_id": {
                    "type": "string"
                },
                "bike_name": {
                    "type": "string"
                },
                "components": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/webike_bike_microservice_nikita_internal_core_domain.Component"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "mileage": {
                    "type": "integer"
                },
                "model": {
                    "description": "stels",
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/webike_bike_microservice_nikita_internal_core_domain.BikeType"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "webike_bike_microservice_nikita_internal_core_domain.BikeType": {
            "type": "string",
            "enum": [
                "bmx",
                "mtb",
                "road"
            ],
            "x-enum-varnames": [
                "BMX",
                "MTB",
                "Road"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}